Move back dump command to CLI plugins
