+++ airflow-core/src/airflow/cli/commands/plugins_command.py
