+++ airflow-core/tests/unit/cli/commands/test_plugins_command.py
@@ -24,7 +24,7 @@ from contextlib import redirect_stdout
 import pytest
 
 from airflow.cli import cli_parser
-from airflow.cli.commands.plugins import dump_command
+from airflow.cli.commands import plugins_command
 from airflow.listeners.listener import get_listener_manager
 from airflow.plugins_manager import AirflowPlugin
 from airflow.sdk import BaseOperatorLink
@@ -38,9 +38,7 @@ pytestmark = pytest.mark.db_test
 class AirflowNewLink(BaseOperatorLink):
     """Operator Link for Apache Airflow Website."""
 
-    @property
-    def name(self) -> str:
-        return "airflowtestlink"
+    name = "airflowtestlink"
 
     def get_link(self, operator, *, ti_key):
         return "https://airflow.apache.org"
@@ -51,7 +49,7 @@ class TestPlugin(AirflowPlugin):
     global_operator_extra_links = [AirflowNewLink()]
 
 
-class TestDumpCommand:
+class TestPluginsCommand:
     @classmethod
     def setup_class(cls):
         cls.parser = cli_parser.get_parser()
@@ -59,14 +57,14 @@ class TestDumpCommand:
     @mock_plugin_manager(plugins=[])
     def test_should_display_no_plugins(self):
         with redirect_stdout(io.StringIO()) as temp_stdout:
-            dump_command.dump_plugins(self.parser.parse_args(["plugins", "dump", "--output=json"]))
+            plugins_command.dump_plugins(self.parser.parse_args(["plugins", "--output=json"]))
             stdout = temp_stdout.getvalue()
         assert "No plugins loaded" in stdout
 
     @mock_plugin_manager(plugins=[ComplexAirflowPlugin])
     def test_should_display_one_plugin(self):
         with redirect_stdout(io.StringIO()) as temp_stdout:
-            dump_command.dump_plugins(self.parser.parse_args(["plugins", "dump", "--output=json"]))
+            plugins_command.dump_plugins(self.parser.parse_args(["plugins", "--output=json"]))
             stdout = temp_stdout.getvalue()
         print(stdout)
         info = json.loads(stdout)
@@ -152,7 +150,7 @@ class TestDumpCommand:
     @mock_plugin_manager(plugins=[TestPlugin])
     def test_should_display_one_plugins_as_table(self):
         with redirect_stdout(io.StringIO()) as temp_stdout:
-            dump_command.dump_plugins(self.parser.parse_args(["plugins", "dump", "--output=table"]))
+            plugins_command.dump_plugins(self.parser.parse_args(["plugins", "--output=table"]))
             stdout = temp_stdout.getvalue()
 
         # Remove leading spaces
@@ -161,8 +159,8 @@ class TestDumpCommand:
         expected_output = textwrap.dedent(
             """\
             name            | global_operator_extra_links
-            ================+====================================================================
-            test-plugin-cli | <unit.cli.commands.plugins.test_dump_command.AirflowNewLink object>
+            ================+===============================================================
+            test-plugin-cli | <unit.cli.commands.test_plugins_command.AirflowNewLink object>
             """
         )
         assert stdout == expected_output
